no code change - fix a comment

+++ android/src/main/java/com/RNFetchBlob/RNFetchBlobFS.java
@@ -607,8 +607,8 @@ class RNFetchBlobFS {
 
     /**
      * List content of folder
-     * @param path Target folder
-     * @param callback  JS context callback
+     * @param path  Target folder
+     * @param promise  JS context promise
      */
     static void ls(String path, Promise promise) {
         try {
@@ -674,7 +674,7 @@ class RNFetchBlobFS {
                 if(read <= 0) {
                     break;
                 }
-                out.write(buffer, 0, (int) Math.min(remain, read));
+                out.write(buffer, 0, Math.min(remain, read));
                 now += read;
             }
             in.close();
